Switch from Pizlo form to basic block arguments
The problem with Pizlo form is that it doesn't seem to work at all well
with SSA-based register allocation.  It works very conveniently with
normal out-of-SSA: I think you can just delete the phis and replace the
upsilons with writes.  However, I cannot work out how to do liveness
analysis on the SSA form in preparation for SSA-based register
allocation.  The lack of connection between Phis and Upsilons and the
fact that Upsilons don't have to appear in direct predecessor blocks to
the blocks that Phis appear in seems to make liveness analysis
impossible to figure out.

On the other hand, I think that basic blocks with arguments is supposed
to be quite tractable for analysis.  It does seem to produce more
verbose output however.

I haven't fixed the lengauer-tarjan file yet.