WIP: Add instruction selection phase
The compiler's design until now was that register allocation happened
from SSA.  That was correct, but it was implemented in terms of the
wrong IR.  The current IR is the architecture-independent IR of the
abstract machine.  The abstract machine does not have registers so doing
register allocation on it is nonsensical.

Instead, the abstract SSA IR needs to be lowered to a machine-specific
SSA IR on which register allocation can be done.  This lowering is
instruction selection.  Each abstract machine operation (ADD, LOAD,
etc.) corresponds to one or more real operations (ADD or ADDI, address
generation + load, etc.).  Abstract machine entities like "variables"
are lowered to real entities like stack slots and registers.

In the machine SSA IR, not everything is lowered to real entities.
Registers and stack slots are still virtual/SSA values.  It is still an
SSA IR and not assembly language.  The difference is that the operations
themselves are real.  Every machine SSA operation corresponds to a real
instruction. (Mostly; every operation corresponds to a real assembly
language (pseudo)instruction, but for example "LA" and "LI" might be
assembled into a series of instructions.)  Instructions are selected
concretely at this stage.  For example, we select ADD or ADDI depending
on whether the argument is an immediate.  On a CISC processor like
x86_64, quite a number of real machine instructions may be selected for
a single abstract machine operations, but that is not implemented yet.

Then when we do register allocation and spilling on the machine SSA we
will get actual assembly language that makes sense: real operations
operating on real entities.