Change procedures to have parameters
Procedures now have parameters and arguments when called.  There is no
type checking yet and just a single type of procedures.  You can set the
return value of the current procedure named 'proc' by assigning to the
variable 'proc'.

In the language, 'call' is now an expression (and so is assignment) and
its value is the return value of the called procedure.

In SSA, call continuations now take a list of parameters.  The
continuation block for the call takes an additonal parameter which is
the return value of the function call that returns to that block.

Register allocation now uses real registers instead of 'r0', 'r1', etc.
but its understanding of those registers is very poor.  If you run out
of registers, the compiler will crash.  It does not understand that it
can use argument registers as temporaries, and it does not understand
the difference between caller-saved and callee-saved registers (so:
procedures do not save any registers (or indeed have any prologue or
epilogue at all) and procedures do not rely on callee-saved registers
being saved.  Much still to do.